+++ src/molecules/checkbox/Checkbox.stories.js
@@ -26,7 +26,7 @@ storiesOf('Molecules/Checkbox', module)
         name="checkbox-default"
         v-model="status"
         value="some value"
-        @input="onChange"
+        @change="onChange"
       >
         Checkbox field
       </a-checkbox>
@@ -45,18 +45,20 @@ storiesOf('Molecules/Checkbox', module)
         input-class="a-checkbox__field"
         v-model="status"
         value="some value"
-        @input="onChange"
+        @change="onChange"
       >
         <template #label>
-          <a-icon
-            icon="checked"
-            class="a-checkbox__icon"
-          />
           <label
             for="checkbox-slots"
             class="a-checkbox__label"
             style="color: #0000ff;"
           >
+            <div class="a-checkbox__icon-wrapper">
+              <a-icon
+                icon="checked"
+                class="a-checkbox__icon"
+              />
+            </div>
             I hereby agree for processing my personal data.
           </label>
         </template>